OAK-9783 - the ExcerptTest is moved to oak-search
